pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jointadaspec"
version = "0.1.0"
description = "Joint Adaptive Speculative Decoding via Tabular MDP"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0",
    "transformers>=4.40",
    "numpy>=1.26",
    "scipy>=1.12",
    "pandas>=2.0",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "tqdm>=4.66",
    "pyarrow>=15.0",
    "scikit-learn>=1.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "mypy>=1.9",
    "types-tqdm",
]

[tool.setuptools.packages.find]
include = ["jointadaspec*", "sp_samp*"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
